fixup!
Since we insist that periods be requested in order, we need to distinguish
whether first period begins at basetime, or that simply anchors when a
period can begin.

We still want to start with invoice request counter 0, so we have a
separate offset field for this case.

The paywindow needs to be more flexible though, and not limited to
when we have a base, so make it a separate field.

The signature is also changed to a standard signature which covers the
entire invoice request rather than a boutique one (why not?).

Examples are put in for period calculations, as they are not simple
(hint: see mktime() for how to implement this, but make sure you set
to UTC first).