fixup!
We can't define period in terms of the previous: it needs to be defined
in terms of the first period.  (e.g. 1 month gives 31 Jan -> 28 Feb -> 31 Mar,
but you need to know that it started on 31 Jan, not 28 Feb!).

Also fix up some whitespace, and insist that there be no recurrence
fields on invreqs for offers without recurrence.