BOLT 12: Change the way "send me money" works.
When I started implementing this, the flow didn't make sense.  If I
just want to send you money, logically it's an "invoice request" but
it has all the fields of an "offer".  So, just use a field in the
offer to indicate "just send an invoice, not an invoice_request".

This means there's *always* an offer for an invoice; the invoice
reflects the `send_invoice` field, to indicate that the description
etc. are reflecting the offer, not the recipient.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>


Header from folded patch 'more-send-invoice.patch':

BOLT 12: reflect `send_invoice` in the invoice message.

This lets you know that this invoice was in response to a send_invoice offer:
vital, since that's where we